requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.1
Authlib==1.3.0
